requirements.txt
viam-sdk==0.43.1
typing-extensions
fasteners